+++ src/controllers/PresentationController.js
@@ -5,34 +5,50 @@ module.exports = {
         const presentations = await connection('presentations')
             .join('groups', 'groups.id', '=', 'presentations.group_id');
 
-        console.log(presentations);
-
-        return response.json(presentations);
+        return response.status(200).json(presentations);
     },
 
     async create(request, response) {
         const { day, time, group_id } = request.body;
 
-        const [id] = await connection('presentations').insert({
-            day,
-            time, 
-            group_id
-        });
+        if(!day || !time || !group_id) {
+            return response.status(400).json({ success: false, message: 'Falta informações na requisição' });
+        }
 
-        return response.json({ id, day, time, group_id });
+        try {
+
+            const [id] = await connection('presentations').insert({
+                day,
+                time, 
+                group_id
+            });
+
+            return response.status(200).json({ id, day, time, group_id });
+        } catch(e) {
+            return response.status(400).json({ success: false, error: e });
+        }
+
     },
 
     async delete(request, response) {
         const { id } = request.params;
 
-        const presentation = await connection('presentations')
-            .where('id', id)
-            .first();
+        try {
 
-
-        await connection('presentations').where('id', id).delete();
-
-        return response.status(204).send();
+            const presentation = await connection('presentations')
+                .where('id', id)
+                .first();
+            
+            if(!presentation) {
+                return response.status(400).json({ success: false, error: e });    
+            }
+                
+            await connection('presentations').where('id', id).delete();
+            
+            return response.status(200).send();
+        } catch(e) {
+            return response.status(400).json({ success: false, error: e });
+        }
 
     }
 }
